Move the enum into its own folder to avoid circual references.
