+++ pygsti/enums/convertspaceenum.py
