+++ pygsti/modelmembers/povms/tensorprodeffect.py
@@ -21,7 +21,7 @@ from pygsti.baseobjs import statespace as _statespace
 from pygsti.tools import listtools as _lt
 from pygsti.tools import matrixtools as _mt
 from pygsti.tools import slicetools as _slct
-from pygsti.tools import SpaceConversionType
+from pygsti.enums import SpaceConversionType
 
 
 class TensorProductPOVMEffect(_POVMEffect):
